+++ base/graph_properties.h
@@ -8,14 +8,14 @@
 #include <base/port.h>
 #include <base/vessel_class.h>
 
-typedef std::unordered_map<std::shared_ptr<Port>, std::pair<float, float>> PortDuals;
-typedef std::unordered_map<std::shared_ptr<VesselClass>, float> VcDuals;
+typedef std::unordered_map<std::shared_ptr<Port>, std::pair<double, double>> PortDuals;
+typedef std::unordered_map<std::shared_ptr<VesselClass>, double> VcDuals;
 
 class GraphProperties {
 public:
     PortDuals                       port_duals;
     PortDuals                       old_port_duals;
-    float                           vc_dual;
+    double                          vc_dual;
     int                             pu_upper_bound;
     int                             de_upper_bound;
     
@@ -23,7 +23,7 @@ public:
     
     GraphProperties(PortDuals port_duals,
                     PortDuals old_port_duals,
-                    float vc_dual,
+                    double vc_dual,
                     int pu_upper_bound,
                     int de_upper_bound) : port_duals(port_duals), old_port_duals(old_port_duals), vc_dual(vc_dual), pu_upper_bound(pu_upper_bound), de_upper_bound(de_upper_bound) {}
 };
