+++ column/column.h
@@ -18,9 +18,9 @@ class Column {
 public:
     std::shared_ptr<const Problem>  prob;
     Solution                        sol;
-    float                           obj_coeff;
-    std::vector<float>              port_coeff;
-    std::vector<float>              vc_coeff;
+    double                          obj_coeff;
+    std::vector<double>              port_coeff;
+    std::vector<double>              vc_coeff;
     bool                            dummy;
     std::string                     created_by;
     ColumnOrigin                    origin; 
@@ -30,14 +30,14 @@ public:
     Column(std::shared_ptr<const Problem> prob, const Solution& sol, const std::string& created_by, ColumnOrigin origin = ColumnOrigin::NONE);
     Column(std::shared_ptr<const Problem> prob,
            const Solution& sol,
-           float obj_coeff,
-           const std::vector<float>& port_coeff,
-           const std::vector<float>& vc_coeff,
+           double obj_coeff,
+           const std::vector<double>& port_coeff,
+           const std::vector<double>& vc_coeff,
            bool dummy,
            const std::string& created_by,
            ColumnOrigin origin) : prob(prob), sol(sol), obj_coeff(obj_coeff), port_coeff(port_coeff), vc_coeff(vc_coeff), dummy(dummy), created_by(created_by), origin(origin) {}
     
-    void make_dummy(float huge_cost);
+    void make_dummy(double huge_cost);
     
     bool is_compatible_with_unite_rule(const VisitRule& vr) const;
     bool is_compatible_with_separate_rule(const VisitRule& vr) const;
