+++ subproblem/heuristics_solver.h
@@ -21,14 +21,14 @@ public:
     HeuristicsSolver(const std::shared_ptr<const Problem> prob, const std::shared_ptr<const Graph> g) : prob(prob), g(g) {}
     
     std::vector<Solution> solve_fast() const;
-    std::vector<Solution> solve_elem_on_reduced_graph(float percentage) const;
-    std::vector<Solution> solve_on_reduced_graph(const float percentage) const { return solve_on_generic_graph(percentage, false); }
+    std::vector<Solution> solve_elem_on_reduced_graph(double percentage) const;
+    std::vector<Solution> solve_on_reduced_graph(const double percentage) const { return solve_on_generic_graph(percentage, false); }
     std::vector<Solution> solve_on_smart_graph() const { return solve_on_generic_graph(0, true); };
     
 private:
     std::vector<Solution> solve_fast_forward() const;
     std::vector<Solution> solve_fast_backward() const;
-    std::vector<Solution> solve_on_generic_graph(float percentage, bool smart) const;
+    std::vector<Solution> solve_on_generic_graph(double percentage, bool smart) const;
 };
 
 #endif
