+++ subproblem/labelling.h
@@ -19,11 +19,11 @@ class Label {
 public:
     int             q_pickupable;
     int             q_deliverable;
-    float           cost;
+    double          cost;
     
     Label(int qp,
           int qd,
-          float c = 0) : q_pickupable(qp), q_deliverable(qd), cost(c) {}
+          double c = 0) : q_pickupable(qp), q_deliverable(qd), cost(c) {}
           
     bool operator==(const Label& other) const;
     bool operator<(const Label& other) const;
@@ -37,7 +37,7 @@ class ElementaryLabel : public Label {
 public:
     VisitedPortsFlags visited_ports;
     
-    ElementaryLabel(int qp, int qd, float c, const VisitedPortsFlags& vp) : Label(qp, qd, c), visited_ports(vp) {}
+    ElementaryLabel(int qp, int qd, double c, const VisitedPortsFlags& vp) : Label(qp, qd, c), visited_ports(vp) {}
     
     bool operator==(const ElementaryLabel& other) const;
     bool operator<(const ElementaryLabel& other) const;
